Adding User Attached entity and form in account management, fixing some bugs

+++ src/AppBundle/Controller/User/AccountController.php
@@ -5,6 +5,7 @@ namespace AppBundle\Controller\User;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
 use AppBundle\Form\Type\UserType;
+use AppBundle\Form\Type\UserSelect;
 use AppBundle\Form\Type\UserNotification;
 use AppBundle\Entity\User;
 use Symfony\Component\HttpFoundation\Request;
@@ -26,18 +27,23 @@ class AccountController extends Controller
 				$delete_form = $this->createForm(UserType::class, $user);
 				$delete_form->handleRequest($request);
 				if ($delete_form->isSubmitted() && $delete_form->isValid()) {
-					if($this->delete($user, $delete_form))
-						return $this->redirectToRoute('homepage');
+					if($delete_form->get('delete')->isClicked()){
+						if($this->delete($user, $delete_form))
+							return $this->redirectToRoute('homepage');
+					}
 				}
 
 				$user_form = $this->createForm(UserType::class, $user);
 				$user_form->handleRequest($request);
 				$notification_form = $this->createForm(UserNotification::class, $user);
 				$notification_form->handleRequest($request);
+				$option_form = $this->createForm(UserSelect::class, $user);
+				$option_form->handleRequest($request);
 				if ($notification_form->isSubmitted() && $notification_form->isValid()
-						|| $user_form->isSubmitted() && $user_form->isValid())
+						|| $user_form->isSubmitted() && $user_form->isValid()
+						|| $option_form->isSubmitted() && $option_form->isValid())
 				{
-					$this->save($user, $user_form, $notification_form);
+					$this->save($user, $user_form, $notification_form, $option_form);
 				}
 
 				return $this->render('user/account/parameter.html.twig', [
@@ -45,6 +51,7 @@ class AccountController extends Controller
 						'form_user' => $user_form->createView(),
 						'delete_form' => $delete_form->createView(),
 						'notification_form' => $notification_form->createView(),
+						'option_form' => $option_form->createView(),
         ]);
     }
 
@@ -52,7 +59,7 @@ class AccountController extends Controller
 		* Function to save an account
 		* @param User $user the user entity to save into database
 		*/
-		private function save(User $user, Form $user_form, Form $notification_form)
+		private function save(User $user, Form $user_form, Form $notification_form, Form $option_form)
 		{
 			if($user->getPlainPassword()){
 				$password = $this->get('security.password_encoder')
